xalloc.h: add new functions: xstrappend, xstrputc, xstrvfappend, and xstrfappend
