+++ misc-utils/lsfd.h
@@ -275,44 +275,6 @@ const char *get_miscdev(unsigned long minor);
 const char *get_nodev_filesystem(unsigned long minor);
 void add_nodev(unsigned long minor, const char *filesystem);
 
-static inline void xstrappend(char **a, const char *b)
-{
-	if (strappend(a, b) < 0)
-		err(XALLOC_EXIT_CODE, _("failed to allocate memory for string"));
-}
-
-static inline void xstrputc(char **a, char c)
-{
-	char b[] = {c, '\0'};
-	xstrappend(a, b);
-}
-
-static inline
-__attribute__((__format__(printf, 2, 0)))
-int xstrvfappend(char **a, const char *format, va_list ap)
-{
-	int ret = strvfappend(a, format, ap);
-
-	if (ret < 0)
-		err(XALLOC_EXIT_CODE, "cannot allocate string");
-	return ret;
-
-}
-
-static inline
-__attribute__ ((__format__ (__printf__, 2, 3)))
-int xstrfappend(char **a, const char *format, ...)
-{
-	va_list ap;
-	int ret;
-
-	va_start(ap, format);
-	ret = xstrvfappend(a, format, ap);
-	va_end(ap);
-
-	return ret;
-}
-
 /*
  * Net namespace
  */
